Update Relationships and Add Exception Returns

+++ src/main/java/com/personal/school/controller/SubjectController.java
@@ -1,6 +1,7 @@
 package com.personal.school.controller;
 
-import com.personal.school.dto.SubjectDto;
+import com.personal.school.dto.SubjectDTO;
+import com.personal.school.exception.NotFoundException;
 import com.personal.school.form.SubjectForm;
 import com.personal.school.model.Subject;
 import com.personal.school.repository.SubjectRepository;
@@ -15,7 +16,7 @@ import java.net.URI;
 import java.util.List;
 import java.util.Optional;
 
-import static com.personal.school.dto.SubjectDto.toDto;
+import static com.personal.school.dto.SubjectDTO.toDto;
 
 @RestController
 @RequestMapping("/subject")
@@ -25,31 +26,36 @@ public class SubjectController {
     SubjectRepository subjectRepository;
 
     @GetMapping
-    public List<SubjectDto> getAllSubject(){
+    public List<SubjectDTO> getAllSubject(){
         List<Subject> subjects = subjectRepository.findAll();
         return toDto(subjects);
     }
 
     @GetMapping
     @RequestMapping("/{id}")
-    public ResponseEntity<SubjectDto> getSubjectById(@PathVariable Long id){
+    public ResponseEntity<SubjectDTO> getSubjectById(@PathVariable Long id){
         Optional<Subject> subject = subjectRepository.findById(id);
-        return subject.map(value -> ResponseEntity.ok(new SubjectDto(value)))
+        return subject.map(value -> ResponseEntity.ok(new SubjectDTO(value)))
                 .orElseGet(() -> ResponseEntity.notFound().build());
     }
 
     @PostMapping
     @Transactional
-    public ResponseEntity<SubjectDto> addSubject(@RequestBody @Valid SubjectForm subjectForm, UriComponentsBuilder uriBuilder){
+    public ResponseEntity<SubjectDTO> addSubject(@RequestBody @Valid SubjectForm subjectForm, UriComponentsBuilder uriBuilder){
         Subject subject = Subject.toSubject(subjectForm);
         subjectRepository.save(subject);
 
         URI uri = uriBuilder.path("/subject/{id}").buildAndExpand(subject.getId()).toUri();
-        return ResponseEntity.created(uri).body(new SubjectDto(subject));
+        return ResponseEntity.created(uri).body(new SubjectDTO(subject));
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> removeSubject(@PathVariable Long id){
+
+        if(!subjectRepository.findById(id).isPresent()) {
+            throw new NotFoundException(String.format("Subject with id [%s] does not exist", id));
+        }
+
         subjectRepository.deleteById(id);
         return ResponseEntity.ok().build();
     }
